requirements.txt
python-telegram-bot==20.7
google-api-python-client
google-auth==2.29.0
google-auth-oauthlib
google-auth-httplib2
gspread==5.12.0
gunicorn==21.2.0
